+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-31 16:04:06
-// Commit: feat(core): add bi-directional event bus handler for core
+// Updated at: 2026-09-01 12:27:10
+// Commit: style(core): adjust design tokens, typography, and contrast in core
 
 class coreComponentState {
-  final String commitRef = '4079';
-  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
-  final DateTime timestamp = DateTime.parse('2026-08-31T16:04:06+05:00');
+  final String commitRef = '4110';
+  final String lastMessage = 'style(core): adjust design tokens, typography, and contrast in core';
+  final DateTime timestamp = DateTime.parse('2026-09-01T12:27:10+05:00');
 }